Выносит в отдельную функцию форматирование даты и времени для комментариев
